# swagger.yaml
swagger: "2.0"
info:
  description: "Este documento é referente a API do Minhas Séries"
  version: "1.0.0"
  title: "Minhas Séries API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "danielcaitano9@hotmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/"
tags:
- name: "serie"
  description: "Tudo sobre séries."
- name: "user"
  description: "Tudo sobre usuários"
schemes:
- "https"
- "http"
paths:
  /series:
    get:
      tags:
      - "serie"
      summary: "Retorna todas as séries"
      produces: 
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items: 
              $ref: "#/definitions/Serie"
      security:
      - api_key:
        - "restrito"

    post:
      tags:
      - "serie"
      summary: "Cria uma nova série"
      description: ""
      operationId: "addSerie"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Série que desejamos criar"
        required: true
        schema:
          $ref: "#/definitions/Serie"
      responses:
        405:
          description: "Invalid input"
      security:
      - api_key:
        - "restrito"
  /series/{serieId}:
    get:
      tags:
      - "serie"
      summary: "procurar série por ID"
      description: "Returns a single pet"
      operationId: "getSerieById"
      produces:
      - "application/json"
      parameters:
      - name: "serieId"
        in: "path"
        description: "ID da série a ser retornada"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Serie"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Pet not found"
      security:
      - api_key: 
        - "restrito"

    delete:
      tags:
      - "serie"
      summary: "Deletes a pet"
      description: ""
      operationId: "deletePet"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "serieId"
        in: "path"
        description: "Pet id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Pet not found"
      security:
      - api_key:
        - "restrito"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "x-access-token"
    in: "header"
    scopes:
      restrito: restrito access
      admin: admin 
definitions:
  Serie:
    type: "object"
    properties: 
      _id: 
        type: "string"
      name:
        type: "string"
      status: 
        type: "string"

  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      userStatus:
        type: "integer"
        format: "int32"
        description: "User Status"
    xml:
      name: "User"